fix: tool without ffmpeg and magick

+++ tool/src/utils/convert_sound.ts
@@ -3,6 +3,9 @@ import fs from "fs/promises";
 import { read_conf } from "../read_conf";
 import { exec_cmd } from "./exec_cmd";
 
+const { FFMPEG_CMD } = read_conf();
+export const is_ffmpeg_exists = command_exists.sync(FFMPEG_CMD)
+export let is_ffmpeg_tried = false ;
 function get_dst_path(
   out_dir: string,
   src_dir: string,
@@ -10,15 +13,24 @@ function get_dst_path(
 ): string {
   return src_path.replace(src_dir, out_dir) + ".mp3";
 }
+export function print_ffmpeg_hints() {
+  if (is_ffmpeg_exists || !is_ffmpeg_tried) return;
+  const hints = `
+ffmpeg not found, download it from: https://ffmpeg.org/download.html
+
+All audio files will not be converted to MP3. 
+However, you can convert them to MP3 in your own way.
+Then you need to put it into output zip file yourself.
+
+'lf2w_tool' currently only supports MP3.
+`
+  console.warn(hints)
+}
 export async function convert_sound(dst_path: string, src_path: string) {
+  is_ffmpeg_tried = true;
+  if (!is_ffmpeg_exists) return;
   console.log("convert", src_path, "=>", dst_path);
   await fs.rm(dst_path, { recursive: true, force: true }).catch(() => void 0);
-  const { FFMPEG_CMD } = await read_conf();
-  if (!command_exists.sync(FFMPEG_CMD))
-    throw new Error(
-      "ffmpeg not found, download it from: https://ffmpeg.org/download.html",
-    );
-
   const args = [
     "-i",
     src_path,
